requirements.txt
opencv-python>=4.8.0
customtkinter>=5.2.0
Pillow>=10.0.0